Safeguard against deprecated DbEdgeRules versions.
Currently we require a DbEdgeRules_vX.json file and it's associated
edge_properties_vX.json files.
If an edge_properties file is not found for a DbEdgeRules file, it
throws an exeption. We should log and entry stating DbEdgeRules and
it's associaded edge_properties files is not found and not throw an
error.

Issue-ID: AAI-21
Change-Id: I05f5b4d14c3cef1586cefe579e38df03ca4fb438
Signed-off-by: Shwetank Dave <shwetank.dave@amdocs.com>